implemented Neural network

+++ src/get_train_test_split/fashion_mnist_data.py
@@ -4,12 +4,12 @@ import numpy as np
 class FashionMnistData:
     def __init__(self, path_to_folder="data"):
         self.train = np.load(f"{path_to_folder}/fashion_train.npy")
-        self.X_train = self.train[:, : -1]
-        self.y_train = self.train[:, -1]
+        self.X_train = self.train[:, : -1].astype(float)
+        self.y_train = self.train[:, -1].astype(int)
 
         self.test = np.load(f"{path_to_folder}/fashion_test.npy")
-        self.X_test = self.test[:, : -1]
-        self.y_test = self.test[:, -1]
+        self.X_test = self.test[:, : -1].astype(float)
+        self.y_test = self.test[:, -1].astype(int)
 
     def get_train_test_split(self, normalize=False):
         if normalize:
